mattter_backend/requirements.txt
Django>=5.0
djangorestframework
django-filter
django-cors-headers
python-dotenv
Pillow
psycopg2-binary
channels
daphne
djangorestframework-gis
